353-SD/CP Favour client token over user token when both available
A bug was found that meant that if a user logged in via the UI and received a token in a cookie
and then navigated to the /docs page and authenticated with a client app, we would favour the
user token in the cookie rather than the client bearer token in the authorisation header.

This lead to a client app being authorised according to the (not yet expired) user access token in the cookies
which could lead to being granted additional or fewer permissions that allowed/required.