328-CP/CD/RP Call Cognito Logout URL
When we log a user out of the UI we only delete the access token cookie.
However a user session remains in Cognito which means that the user can simply click Login again and be authenticated.

This presents a security risk.

This commit forces a logout of the user session in Cognito.