000-SD/CP Create a single schema for each S3 path
There was a bug that caused the Crawler to create a serparate table for each partition.

This happened when:
- A dataset with a string column containing null values was used
- The dataset was partitioned on another column
- At least one of the partitions contained only values from the string column that were null

This happens because the Crawler infers the schema per partition and so sets a parititon
with all null values to type 'int' even though the overarching column is of type 'string'.

For example, in the dataset below:
- Parititon on 'location'
- 'name' column is type string, but contains null values
- The 'name' column values for  partition 'city2' are all null

location,name,surname
city1,name1,surname1
city1,name2,surname1
city1,name3,surname1
city2,,surname1
city2,,surname1

Forcing the crawler to use a single schema per S3 path, solves this problem.